+++ contracts/non-fungible-tokens/andromeda-cw721-timelock/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "andromeda-nft-timelock"
+name = "andromeda-cw721-timelock"
 version = "0.1.0"
 edition = "2018"
