+++ packages/andromeda-non-fungible-tokens/src/lib.rs
@@ -3,6 +3,6 @@ pub mod crowdfund;
 pub mod cw721;
 pub mod cw721_offers;
 pub mod cw721_staking;
+pub mod cw721_timelock;
 pub mod gumball;
-pub mod nft_timelock;
 pub mod wrapped_cw721;
